feat:FAB 변경, 카테고리 검색결과 화면 추가
