+++ app/src/main/java/com/example/mapin/ui/main_content_detail/MainContentDetailFragment.kt
@@ -40,7 +40,6 @@ class MainContentDetailFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         viewModel = ViewModelProvider(this)[MainContentDetailViewModel::class.java]
-        (activity as MainActivity).findViewById<FloatingActionButton>(R.id.fab).visibility =View.INVISIBLE
         val mapView = MapView(requireContext())
         mapView.setMapCenterPoint(MapPoint.mapPointWithGeoCoord(37.24793398172821, 127.0764451494671), false)
 
@@ -63,10 +62,7 @@ class MainContentDetailFragment : Fragment() {
         }
     }
 
-    //FAB이 MainActivity에 있어서 임시로 해놓은 기능.
-    //기존의 경우 MainContentDetailFragment에서 뒤로가면 FAB이 안보이기 때문에
     override fun onDestroyView() {
-        (activity as MainActivity).findViewById<FloatingActionButton>(R.id.fab).visibility =View.VISIBLE
         super.onDestroyView()
         _binding = null
     }
