+++ app/src/main/res/layout/activity_main.xml
@@ -15,191 +15,6 @@
     <include layout="@layout/content_main" />
 
 
-    <View
-        android:id="@+id/fabBGLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#74000000"
-        android:visibility="gone" />
-
-    <LinearLayout
-        android:id="@+id/fabLayout1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="16dp"
-        android:clipToPadding="false"
-        android:gravity="center_vertical"
-        android:visibility="gone">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/fab_menu1"
-            android:textStyle="bold"
-            android:textColor="@color/white"/>
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            app:srcCompat="@drawable/baseline_storefront_24"
-            app:tint="@android:color/white"
-            app:fabCustomSize="35dp"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/fabLayout2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="16dp"
-        android:gravity="center_vertical"
-        android:visibility="gone">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/fab_menu2"
-            android:textStyle="bold"
-            android:textColor="@color/white"/>
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            app:srcCompat="@drawable/baseline_location_on_24"
-            app:tint="@android:color/white"
-            app:fabCustomSize="35dp"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
-
-    <LinearLayout
-        android:id="@+id/fabLayout3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="16dp"
-        android:clipToPadding="false"
-        android:gravity="center_vertical"
-        android:visibility="gone">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/fab_menu3"
-            android:textStyle="bold"
-            android:textColor="@color/white"/>
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab3"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            app:srcCompat="@drawable/baseline_category_24"
-            app:tint="@android:color/white"
-            app:fabCustomSize="35dp"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/fabLayout4"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="16dp"
-        android:clipToPadding="false"
-        android:gravity="center_vertical"
-        android:visibility="gone">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/fab_menu4"
-            android:textStyle="bold"
-            android:textColor="@color/white"/>
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            app:srcCompat="@drawable/baseline_create_24"
-            app:tint="@android:color/white"
-            app:fabCustomSize="35dp"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/fabLayout5"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="16dp"
-        android:clipToPadding="false"
-        android:gravity="center_vertical"
-        android:visibility="gone">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/fab_menu5"
-            android:textStyle="bold"
-            android:textColor="@color/white"/>
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab5"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            app:srcCompat="@drawable/baseline_post_add_24"
-            app:tint="@android:color/white"
-            app:fabCustomSize="35dp"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:clipToPadding="false"
-        android:gravity="center_vertical">
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="bottom|end"
-            android:layout_margin="@dimen/fab_margin"
-            android:gravity="center_vertical"
-            app:srcCompat="@drawable/baseline_add_24"
-            app:tint="@android:color/white"
-            app:fabSize="normal"
-            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            android:backgroundTint="@color/navy"
-            app:borderWidth="0dp"/>
-    </LinearLayout>
-
 
 
 </FrameLayout>
