+++ app/src/main/res/layout/fragment_search_category.xml
@@ -13,7 +13,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
         android:layout_marginStart="12dp"
-        android:text="카테고리 선택"
+        android:text="카테고리 검색"
         android:textColor="@color/black"
         android:textSize="20sp"
         android:textStyle="bold"
